+++ src/main.py
@@ -1,11 +1,21 @@
 import pygame
-
+import mechanics.card_actions as actions
+import mechanics.card_logic as logic
+from ui.main_menu import MainMenuScene
+from mechanics.scenemanager import Scenemanager
 def main():
     # pygame setup
     pygame.init()
     screen = pygame.display.set_mode((1280, 720))
     clock = pygame.time.Clock()
     running = True
+
+    reader = actions.ActionLoader()
+    pool = logic.CardPool(reader)
+    print(str(pool.cardStats['GUARD']))
+    mainmenu= MainMenuScene("main-menu")
+    sceneManager = Scenemanager([mainmenu],screen)
+    sceneManager.set_active_scene("main-menu")
 
     while running:
         # poll for events
@@ -16,13 +26,13 @@ def main():
 
         # fill the screen with a color to wipe away anything from last frame
         screen.fill("purple")
-
+        sceneManager.update()
         # RENDER YOUR GAME HERE
 
         # flip() the display to put your work on screen
         pygame.display.flip()
 
     clock.tick(60)
-
-
-main()
+    
+if __name__ == "__main__":
+    main()