+++ src/ui/scene.py
@@ -1,0 +1,12 @@
+class Scene:
+    def __init__(self,name,scenemanager = None):
+        self.name = name
+        if scenemanager != None:
+            scenemanager.add_scene(self)
+        self.scenemanager = scenemanager
+
+    def update(self):
+        pass
+    
+    def get_scene_name(self):
+        return self.name