+++ Documentation/changelog.md
@@ -1,3 +1,10 @@
 ## Week 3
 - Made a general drawpile class to handle drawing and shufflig the deck
 - Tests for the drawpile have been added
+
+## Week 4
+- Card data is now loaded from a file
+- Main menu added
+- Scenemanagement added
+- Card Interactions have been added
+- Majority core features (health,card-logic etc..) have been added but not yet implemented with oneother